move ProfileEvent to api.bukkit.event package
