+++ src/main/java/org/betonquest/betonquest/api/Objective.java
@@ -3,8 +3,8 @@ package org.betonquest.betonquest.api;
 import org.betonquest.betonquest.BetonQuest;
 import org.betonquest.betonquest.GlobalObjectives;
 import org.betonquest.betonquest.Instruction;
-import org.betonquest.betonquest.api.bukkit.events.PlayerObjectiveChangeEvent;
-import org.betonquest.betonquest.api.bukkit.events.QuestDataUpdateEvent;
+import org.betonquest.betonquest.api.bukkit.event.PlayerObjectiveChangeEvent;
+import org.betonquest.betonquest.api.bukkit.event.QuestDataUpdateEvent;
 import org.betonquest.betonquest.api.logger.BetonQuestLogger;
 import org.betonquest.betonquest.api.profiles.OnlineProfile;
 import org.betonquest.betonquest.api.profiles.Profile;
