+++ src/main/java/org/betonquest/betonquest/api/bukkit/event/LoadDataEvent.java
@@ -1,4 +1,4 @@
-package org.betonquest.betonquest.api.bukkit.events;
+package org.betonquest.betonquest.api.bukkit.event;
 
 import org.bukkit.event.Event;
 import org.bukkit.event.HandlerList;
