+++ src/main/java/org/betonquest/betonquest/api/bukkit/event/PlayerTagRemoveEvent.java
@@ -1,7 +1,6 @@
-package org.betonquest.betonquest.api.bukkit.events;
+package org.betonquest.betonquest.api.bukkit.event;
 
 import org.betonquest.betonquest.api.profiles.Profile;
-import org.betonquest.betonquest.api.profiles.ProfileEvent;
 import org.bukkit.event.HandlerList;
 
 /**
