+++ src/main/java/org/betonquest/betonquest/compatibility/citizens/CitizensWalkingListener.java
@@ -5,8 +5,8 @@ import net.citizensnpcs.api.event.SpawnReason;
 import net.citizensnpcs.api.npc.NPC;
 import org.apache.commons.lang3.tuple.Pair;
 import org.betonquest.betonquest.BetonQuest;
-import org.betonquest.betonquest.api.bukkit.events.PlayerConversationEndEvent;
-import org.betonquest.betonquest.api.bukkit.events.PlayerConversationStartEvent;
+import org.betonquest.betonquest.api.bukkit.event.PlayerConversationEndEvent;
+import org.betonquest.betonquest.api.bukkit.event.PlayerConversationStartEvent;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.event.EventHandler;
